pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "^1.4.37"
aiomysql = "^0.1.1"
aioredis = "^2.0.1"
hiredis = "^2.0.0"
redis = "^4.3.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.9.1"}

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.18.3"
aiosqlite = "^0.17.0"
httpx = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = 'strict'